Fix offset macros for 64-bit compatibility using offsetof

+++ src/bgame/Engine.h
@@ -11,7 +11,7 @@
 class Engine
 {
 public:
-    typedef int (QDECL* Ptr)( int, ... );
+    typedef intptr_t (QDECL* Ptr)( intptr_t, ... );
 
     static Ptr ptr;
